+++ test/clj/rems/api/test_applications.clj
@@ -483,11 +483,13 @@
       (let [created (get-application-for-user application-id user-id)]
         (is (= "application.state/draft" (get created :application/state)))))
 
-    (testing "getting application as other user is forbidden"
-      (is (response-is-forbidden?
-           (-> (request :get (str "/api/applications/" application-id))
-               (authenticate api-key "bob")
-               handler))))
+    (testing "seeing draft is forbidden"
+      (testing "as unrelated user"
+        (is (response-is-forbidden? (api-response :get (str "/api/applications/" application-id) nil api-key "bob"))))
+      (testing "as reporter"
+        (is (response-is-forbidden? (api-response :get (str "/api/applications/" application-id) nil api-key "reporter"))))
+      (testing "as handler"
+        (is (response-is-forbidden? (api-response :get (str "/api/applications/" application-id) nil api-key "developer")))))
 
     (testing "modifying application as other user is forbidden"
       (is (= {:success false
@@ -503,7 +505,13 @@
         (is (= "application.state/submitted" (get submitted :application/state)))
         (is (= ["application.event/created"
                 "application.event/submitted"]
-               (map :event/type (get submitted :application/events))))))))
+               (map :event/type (get submitted :application/events))))))
+
+    (testing "seeing submitted application as reporter is allowed"
+      (is (response-is-ok?
+            (-> (request :get (str "/api/applications/" application-id))
+                (authenticate api-key "reporter")
+                handler))))))
 
 (deftest test-application-delete
   (let [api-key "42"
@@ -511,10 +519,6 @@
         handler "developer"]
     (let [app-id (test-data/create-application! {:actor applicant})]
       (testing "can't delete draft as other user"
-        (is (not (contains? (-> (get-application-for-user app-id handler)
-                                :application/permissions
-                                set)
-                            :application.command/delete)))
         (is (= {:errors [{:type "forbidden"}] :success false}
                (api-call :post "/api/applications/delete" {:application-id app-id}
                          api-key handler))))